my_flask_project/pyproject.toml
[project]
name = "my-flask-project"
version = "0.1.0"
description = ""
authors = [
    {name = "Білецька",email = "khrista.biletcka7@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "flask (>=3.1.3,<4.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "jikanpy-v4 (>=1.0.2,<2.0.0)"
]

[tool.poetry]
packages = [{include = "my_flask_project", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
